move files to process
