+++ packages/cli/src/tokens/build.ts
@@ -6,15 +6,15 @@ import * as R from 'ramda';
 import StyleDictionary from 'style-dictionary';
 
 import { cleanDir, readFile, writeFile } from '../utils.js';
-import { configs, getConfigsForThemeDimensions } from './build/configs.js';
+import { configs, getConfigsForThemeDimensions } from './process/configs.js';
 import {
   type BuildConfig,
   type SDConfigForThemePermutation,
   type ThemePermutation,
   colorCategories,
-} from './build/types.js';
-import { concatFiles } from './build/utils/entryfile.js';
-import { type ProcessedThemeObject, processThemeObject } from './build/utils/getMultidimensionalThemes.js';
+} from './process/types.js';
+import { concatFiles } from './process/utils/entryfile.js';
+import { type ProcessedThemeObject, processThemeObject } from './process/utils/getMultidimensionalThemes.js';
 import type { TokensSet } from './types.js';
 
 type BuildOptions = {
